test/topology: split cluster features test
Split the test cases into individual files as each takes minutes in
debug mode and combined it piled up to 19 minutes.
